+++ sf-logging/src/main/java/io/sunflower/logging/FileAppenderFactory.java
@@ -103,6 +103,16 @@ import javax.validation.constraints.Min;
  * the default of 8KB to 256KB is reported to significantly reduce thread contention.
  * </td>
  * </tr>
+ * <tr>
+ * <td>{@code immediateFlush}</td>
+ * <td>{@code true}</td>
+ * <td>
+ * If set to true, log events will be immediately flushed to disk. Immediate flushing is safer, but
+ * it degrades logging throughput.
+ * See <a href="https://logback.qos.ch/manual/appenders.html#immediateFlush">the Logback documentation</a>
+ * for details.
+ * </td>
+ * </tr>
  * </table>
  *
  * @author michael
@@ -125,6 +135,17 @@ public class FileAppenderFactory<E extends DeferredProcessingAware> extends
 
     @MinSize(1)
     private Size bufferSize = Size.bytes(FileAppender.DEFAULT_BUFFER_SIZE);
+
+    private boolean immediateFlush = true;
+
+    public boolean isImmediateFlush() {
+        return immediateFlush;
+    }
+
+    @JsonProperty
+    public void setImmediateFlush(boolean immediateFlush) {
+        this.immediateFlush = immediateFlush;
+    }
 
     @JsonProperty
     public String getCurrentLogFilename() {
@@ -227,6 +248,7 @@ public class FileAppenderFactory<E extends DeferredProcessingAware> extends
         layoutEncoder.setLayout(buildLayout(context, layoutFactory));
         appender.setEncoder(layoutEncoder);
 
+        appender.setImmediateFlush(isImmediateFlush());
         appender.setPrudent(false);
         appender.addFilter(levelFilterFactory.build(threshold));
         getFilterFactories().forEach(f -> appender.addFilter(f.build()));