+++ sf-util/src/main/java/io/sunflower/util/Generics.java
@@ -62,11 +62,18 @@ public class Generics {
                             }
                         }
                     }
+                } else if (param instanceof ParameterizedType) {
+                    final Type rawType = ((ParameterizedType) param).getRawType();
+                    if (rawType instanceof Class<?>) {
+                        final Class<T> cls = determineClass(bound, rawType);
+                        if (cls != null) {
+                            return cls;
+                        }
+                    }
                 }
             }
         }
-        throw new IllegalStateException(
-                "Cannot figure out type parameterization for " + klass.getName());
+        throw new IllegalStateException("Cannot figure out type parameterization for " + klass.getName());
     }
 
     @SuppressWarnings("unchecked")